refactor: ARCH-006 merge Shallow sim/components.rs
- Move WallSurface from sim/components.rs to sim/construction.rs
- Update all imports to use sim::construction::WallSurface
- Remove pub mod components from sim/mod.rs
- Delete sim/components.rs

Closes #626
